+++ controller/callbackcontroller.php
@@ -548,7 +548,8 @@ class CallbackController extends Controller {
 
                     if (!$isForcesave
                         && !$prevIsForcesave
-                        && $this->versionManager->available) {
+                        && $this->versionManager->available
+                        && $this->config->GetVersionHistory()) {
                         $changes = null;
                         if (!empty($changesurl)) {
                             $changesurl = $this->config->ReplaceDocumentServerUrlToInternal($changesurl);
@@ -557,7 +558,7 @@ class CallbackController extends Controller {
                         FileVersions::saveHistory($file->getFileInfo(), $history, $changes, $prevVersion);
                     }
 
-                    if (!empty($user)) {
+                    if (!empty($user) && $this->config->GetVersionHistory()) {
                         FileVersions::saveAuthor($file->getFileInfo(), $user);
                     }