+++ js/settings.js
@@ -124,6 +124,7 @@
 
             var sameTab = $("#onlyofficeSameTab").is(":checked");
             var preview = $("#onlyofficePreview").is(":checked");
+            var versionHistory = $("#onlyofficeVersionHistory").is(":checked");
 
             var limitGroupsString = $("#onlyofficeGroups").prop("checked") ? $("#onlyofficeLimitGroups").val() : "";
             var limitGroups = limitGroupsString ? limitGroupsString.split("|") : [];
@@ -144,6 +145,7 @@
                     editFormats: editFormats,
                     sameTab: sameTab,
                     preview: preview,
+                    versionHistory: versionHistory,
                     limitGroups: limitGroups,
                     chat: chat,
                     compactHeader: compactHeader,
@@ -178,6 +180,24 @@
             } else {
                 $("#onlyofficeSecret").attr("type", "password");
             }
+        });
+
+        $("#onlyofficeClearVersionHistory").click(function () {
+            $(".section-onlyoffice").addClass("icon-loading");
+
+            $.ajax({
+                method: "DELETE",
+                url: OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/settings/history"),
+                success: function onSuccess(response) {
+                    $(".section-onlyoffice").removeClass("icon-loading");
+                    if (response) {
+                        var message = t(OCA.Onlyoffice.AppName, "All history successfully deleted");
+                        OC.Notification.show(message, {
+                            timeout: 3
+                        });
+                    }
+                }
+            });
         });
     });